add auth-hook
